Refactor reservation unit info section
