+++ apps/ui/public/locales/en/reservationUnit.json
@@ -36,18 +36,21 @@
   "reservationUpdateFailed": "Booking update failed",
   "priceFree": "Free",
   "relatedReservationUnits": "Also available at this location",
-  "reservationInfo1-1": "You can make bookings <bold>between {{reservationsMaxDaysBefore}} {{unit}}</bold> and <bold>{{reservationsMinDaysBefore}} days in advance</bold>.",
-  "reservationInfo1-2": "You can make bookings <bold> {{reservationsMaxDaysBefore}} {{unit}} in advance</bold>.",
-  "reservationInfo1-3": "You can make bookings <bold>{{reservationsMinDaysBefore}} days in advance</bold>.",
-  "reservationInfo2": "<bold>This item will be available until {{reservationEnds}}</bold>.",
-  "reservationInfo2-1": "<bold>This item will be available {{date}} at {{time}}</bold>.",
-  "reservationInfo2-2": "<bold>This item will be available until {{date}} at {{time}}</bold>.",
-  "reservationInfo2-3": "<bold>Booking of this item has ended {{date}} at {{time}}</bold>.",
-  "reservationInfo3": "The duration of the booking should be between <bold>{{minReservationDuration}}</bold> – <bold>{{maxReservationDuration}}</bold>.",
-  "reservationInfo4_one": "You can have <bold>up to one booking at a time</bold>.",
-  "reservationInfo4_other": "You can have <bold>up to {{count}} bookings at a time</bold>.",
-  "reservationInfo1-weeks": "weeks",
-  "reservationInfo1-months": "months",
+  "reservationInfoSection": {
+    "dayLimitPrefix": "You can make bookings",
+    "dayLimitMax": "earliest {{days}} in advance",
+    "dayLimitMin": "before {{days}} days in advance",
+    "willOpen": "This item will be available {{dateTime}}.",
+    "isOpen": "This item will be available until {{dateTime}}.",
+    "hasClosed": "Booking of this item has ended {{dateTime}}.",
+    "duration": "The duration of the booking should be between ",
+    "maxReservationsPerUserPrefix": "You can have",
+    "maxReservationsPerUser_one": "only one booking at a time",
+    "maxReservationsPerUser_other": "up to {{count}} bookings at a time",
+    "days": "days",
+    "weeks": "weeks",
+    "months": "months"
+  },
   "futurePricingNotice": "<bold>The price will change on {{date}}. The new price is {{price}}</bold>.",
   "futurePriceNoticeTax": "(incl. VAT {{tax}} %)",
   "notifications": {
