+++ apps/ui/public/locales/fi/reservationUnit.json
@@ -36,18 +36,21 @@
   "reservationUpdateFailed": "Varauksen päivitys epäonnistui",
   "priceFree": "Maksuton",
   "relatedReservationUnits": "Myös varattavissa tässä toimipisteessä",
-  "reservationInfo1-1": "Voit tehdä varauksen <bold>aikaisintaan {{reservationsMaxDaysBefore}} {{unit}}</bold> ja <bold>viimeistään {{reservationsMinDaysBefore}} päivää etukäteen</bold>.",
-  "reservationInfo1-2": "Voit tehdä varauksen <bold>aikaisintaan {{reservationsMaxDaysBefore}} {{unit}} etukäteen</bold>.",
-  "reservationInfo1-3": "Voit tehdä varauksen <bold>viimeistään {{reservationsMinDaysBefore}} päivää etukäteen</bold>.",
-  "reservationInfo2": "<bold>Tämä kohde on varattavissa {{reservationEnds}}</bold> asti.",
-  "reservationInfo2-1": "<bold>Tämä kohde on varattavissa {{date}} klo {{time}}</bold>.",
-  "reservationInfo2-2": "<bold>Tämä kohde on varattavissa {{date}} klo {{time}}</bold> asti.",
-  "reservationInfo2-3": "<bold>Tämän kohteen varaaminen on sulkeutunut {{date}} klo {{time}}</bold>.",
-  "reservationInfo3": "Varauksen keston tulee olla välillä <bold>{{minReservationDuration}}</bold> - <bold>{{maxReservationDuration}}</bold>.",
-  "reservationInfo4_one": "Sinulla voi olla samanaikaisesti <bold>enintään yksi varaus</bold>.",
-  "reservationInfo4_other": "Sinulla voi olla samanaikaisesti <bold>enintään {{count}} varausta</bold>.",
-  "reservationInfo1-weeks": "viikkoa",
-  "reservationInfo1-months": "kuukautta",
+  "reservationInfoSection": {
+    "dayLimitPrefix": "Voit tehdä varauksen",
+    "dayLimitMax": "aikaisintaan {{days}} etukäteen",
+    "dayLimitMin": "viimeistään {{days}} päivää etukäteen",
+    "willOpen": "Tämä kohde on varattavissa {{dateTime}}.",
+    "isOpen": "Tämä kohde on varattavissa {{dateTime}} asti.",
+    "hasClosed": "Tämän kohteen varaaminen on sulkeutunut {{dateTime}}.",
+    "duration": "Varauksen keston tulee olla välillä",
+    "maxReservationsPerUserPrefix": "Sinulla voi olla samanaikaisesti",
+    "maxReservationsPerUser_one": "vain yksi varaus",
+    "maxReservationsPerUser_other": "enintään {{count}} varausta",
+    "days": "päivää",
+    "weeks": "viikkoa",
+    "months": "kuukautta"
+  },
   "futurePricingNotice": "<bold>Tämän kohteen hinta muuttuu {{date}} alkaen. Uusi hinta on {{price}}</bold>",
   "futurePriceNoticeTax": " (sis. alv. {{tax}}%)",
   "notifications": {
