+++ apps/ui/public/locales/sv/reservationUnit.json
@@ -36,18 +36,21 @@
   "reservationUpdateFailed": "Uppdateringen misslyckades",
   "priceFree": "Avgiftsfri",
   "relatedReservationUnits": "Kan också bokas vid detta verksamhetsställe",
-  "reservationInfo1-1": "Du kan boka <bold>tidigast {{reservationsMaxDaysBefore}} {{unit}}</bold> och <bold>senast {{reservationsMinDaysBefore}} dagar på förhand</bold>.",
-  "reservationInfo1-2": "Du kan boka <bold>tidigast {{reservationsMaxDaysBefore}} {{unit}} på förhand</bold>.",
-  "reservationInfo1-3": "Du kan boka <bold>senast {{reservationsMinDaysBefore}} dagar på förhand</bold>.",
-  "reservationInfo2": "<bold>Detta objekt kan bokas fram till {{reservationEnds}}</bold>.",
-  "reservationInfo2-1": "<bold>Detta objekt kan bokas fram {{date}} kl. {{time}}</bold>.",
-  "reservationInfo2-2": "<bold>Detta objekt kan bokas fram till {{date}} kl. {{time}}</bold>.",
-  "reservationInfo2-3": "<bold>Bokningen av detta objekt har slutat {{date}} kl. {{time}}</bold>.",
-  "reservationInfo3": "Bokningens längd ska vara mellan <bold>{{minReservationDuration}}</bold> – <bold>{{maxReservationDuration}}</bold>.",
-  "reservationInfo4_one": "Du kan ha <bold>högst en bokning i taget</bold>.",
-  "reservationInfo4_other": "Du kan ha <bold>högst {{count}} bokningar i taget</bold>.",
-  "reservationInfo1-weeks": "veckor",
-  "reservationInfo1-months": "månader",
+  "reservationInfoSection": {
+    "dayLimitPrefix": "Du kan boka",
+    "dayLimitMax": "tidigast {{days}} på förhand",
+    "dayLimitMin": "senast {{days}} dagar på förhand",
+    "willOpen": "Detta objekt kan bokas fram {{dateTime}}.",
+    "isOpen": "Detta objekt kan bokas fram till {{dateTime}}.",
+    "hasClosed": "Bokningen av detta objekt har slutat {{dateTime}}.",
+    "duration": "Bokningens längd ska vara mellan",
+    "maxReservationsPerUserPrefix": "Du kan ha",
+    "maxReservationsPerUser_one": "högst en bokning i taget",
+    "maxReservationsPerUser_other": "högst {{count}} bokningar i taget",
+    "days": "dagar",
+    "weeks": "veckor",
+    "months": "månader"
+  },
   "futurePricingNotice": "<bold>Priset ändras från och med {{date}}. Det nya priset är {{price}}</bold>.",
   "futurePriceNoticeTax": "(inkl. moms {{tax}} %)",
   "notifications": {
